+++ tp1/src/msort.h
@@ -4,6 +4,6 @@
 #include <string.h>
 
 void msort(void *base, size_t num, size_t size,
-           int (*compar)(const void *, const void *));
+           long long (*compar)(const void *, const void *));
 
 #endif
